feat: 重构资源管理，统一CACHE_DIR的引用，优化代码结构
