+++ llm_web_kit/model/html_layout_cls.py
@@ -4,10 +4,10 @@ from typing import List, Union
 from llm_web_kit.config.cfg_reader import load_config
 from llm_web_kit.libs.logger import mylogger as logger
 from llm_web_kit.model.html_classify.model import Markuplm
-from llm_web_kit.model.resource_utils.download_assets import (
-    CACHE_DIR, download_auto_file)
+from llm_web_kit.model.resource_utils.download_assets import download_auto_file
 from llm_web_kit.model.resource_utils.unzip_ext import (get_unzip_dir,
                                                         unzip_local_file)
+from llm_web_kit.model.resource_utils.utils import CACHE_DIR
 
 
 class HTMLLayoutClassifier:
