+++ llm_web_kit/model/quality_model.py
@@ -19,10 +19,10 @@ from llm_web_kit.model.basic_functions.features import (
     stats_html_entity, stats_ngram_mini, stats_punctuation_end_sentence,
     stats_stop_words, stats_unicode)
 from llm_web_kit.model.basic_functions.utils import div_zero
-from llm_web_kit.model.resource_utils.download_assets import (
-    CACHE_DIR, download_auto_file)
+from llm_web_kit.model.resource_utils.download_assets import download_auto_file
 from llm_web_kit.model.resource_utils.unzip_ext import (get_unzip_dir,
                                                         unzip_local_file)
+from llm_web_kit.model.resource_utils.utils import CACHE_DIR
 
 _global_quality_model = {}
 _model_resource_map = {
