+++ llm_web_kit/model/unsafe_words_detector.py
@@ -10,10 +10,10 @@ from llm_web_kit.input.datajson import DataJson
 from llm_web_kit.libs.standard_utils import json_loads
 from llm_web_kit.model.basic_functions.format_check import (is_en_letter,
                                                             is_pure_en_word)
-from llm_web_kit.model.resource_utils.download_assets import (
-    CACHE_DIR, download_auto_file)
+from llm_web_kit.model.resource_utils.download_assets import download_auto_file
 from llm_web_kit.model.resource_utils.singleton_resource_manager import \
     singleton_resource_manager
+from llm_web_kit.model.resource_utils.utils import CACHE_DIR
 
 xyz_language_lst = [
     'ar',
